+++ teddecor/__init__.py
@@ -5,7 +5,8 @@ This is a easy to use library that gives a user access to colored text, bold tex
 underlined text, hyperlinks and much more.
 """
 
-__version__ = "1.1.2"
+__version__ = "1.2.0"
 from .TED import TED
 from .logger import LL, Log, Logger
-from .pprint import p_value
+from .pprint import p_value, pprint
+from . import decorators